extract and memoise activity table to improve performance
