+++ visitor-app/src/cb_admin/pages/Activities/index.js
@@ -1,37 +1,23 @@
 import React from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import { assocPath, dissoc, invertObj } from 'ramda';
+import { assocPath, dissoc } from 'ramda';
 import { Grid, Row, Col } from 'react-flexbox-grid';
 import { FlexContainerCol } from '../../../shared/components/layout/base';
 import { Paragraph, ErrorParagraph } from '../../../shared/components/text/base';
 import { Form, PrimaryButton } from '../../../shared/components/form/base';
 import LabelledInput from '../../../shared/components/form/LabelledInput';
 import LabelledSelect from '../../../shared/components/form/LabelledSelect';
-import _Checkbox from '../../../shared/components/form/StyledLabelledCheckbox';
 import NavHeader from '../../../shared/components/NavHeader';
 import { Activities, CommunityBusiness, ErrorUtils } from '../../../api';
 import { redirectOnError } from '../../../util';
-import ActivityLabel from './ActivityLabel';
-import CategorySelect from './CategorySelect';
-import { colors } from '../../../shared/style_guide';
+import ActivityTable from './ActivityTable';
 
 
 const SubmitButton = styled(PrimaryButton)`
   margin-left: 2em;
   height: 3em;
 `;
-
-
-const Checkbox = styled(_Checkbox)`
-  margin: 0 auto;
-  width: 2rem;
-
-  & input + label::before {
-    border: 0.1rem solid ${colors.light}
-  }
-`;
-
 
 const ActivitiesError = styled(ErrorParagraph)`
   opacity: ${props => (props.visible ? '1' : '0')};
@@ -41,37 +27,6 @@ const ActivitiesError = styled(ErrorParagraph)`
   transition: opacity 0.7s ease;
 `;
 
-const Table = styled.table`
-  background: transparent;
-  width: 100%;
-  padding: 2em;
-  table-layout: fixed;
-`;
-const TableHead = styled.thead``;
-const TableBody = styled.tbody``;
-const TableRow = styled.tr`
-  height: 3em;
-`;
-const TableCell = styled.td.attrs(props => ({ colSpan: props.wide ? 5 : 1 }))`
-  text-align: ${props => (props.center ? 'center' : 'left')};
-  vertical-align: middle;
-`;
-const TableHeader = styled.th.attrs(props => ({ colSpan: props.wide ? 5 : 1 }))``;
-
-
-const keyMap = {
-  name: 'Activity',
-  category: 'Category',
-  monday: 'Mon',
-  tuesday: 'Tue',
-  wednesday: 'Wed',
-  thursday: 'Thu',
-  friday: 'Fri',
-  saturday: 'Sat',
-  sunday: 'Sun',
-};
-const colToState = invertObj(keyMap);
-const columns = Object.values(keyMap);
 
 const doesActivityAlreadyExist = (newActivityName, activitiesObject) =>
   Object.values(activitiesObject).some(({ name }) =>
@@ -150,7 +105,6 @@ export default class ActivitiesPage extends React.Component {
       );
   }
 
-
   deleteActivity = (id) => {
     Activities.delete({ id })
       .then(() => {
@@ -218,59 +172,13 @@ export default class ActivitiesPage extends React.Component {
         <ActivitiesError visible={errors.view}>
           {errors.general}
         </ActivitiesError>
-        <Table>
-          <TableHead>
-            <TableRow>
-              {
-                columns.map((col, i) => <TableHeader key={col} wide={i < 2}>{col}</TableHeader>)
-              }
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {
-              this.state.activities.order.map((id) => {
-                const activity = this.state.activities.items[id];
-                const k1 = colToState[columns[0]];
-                const k2 = colToState[columns[1]];
-
-                return (
-                  <TableRow key={activity.id}>
-                    <TableCell key={`${activity[k1]}-${k1}`} wide>
-                      <ActivityLabel
-                        label={activity[k1]}
-                        onClick={() => this.deleteActivity(activity.id)}
-                      />
-                    </TableCell>
-                    <TableCell key={`${activity[k2]}-${k2}`} center wide>
-                      <CategorySelect
-                        id={id}
-                        options={this.state.categories}
-                        value={activity.category}
-                        onChange={e => this.updateCategory(activity.id, e)}
-                      />
-                    </TableCell>
-                    {
-                      columns
-                        .slice(2)
-                        .map(k => colToState[k])
-                        .map(k => (
-                          <TableCell key={`${activity[k]}-${k}`} center>
-                            <Checkbox
-                              id={`${activity.id}-${k}`}
-                              name={`${activity.id}-${k}`}
-                              alt={`${activity.name} ${k} update button`}
-                              checked={activity[k]}
-                              onChange={() => this.toggleCheckbox(activity.id, k)}
-                            />
-                          </TableCell>
-                        ))
-                    }
-                  </TableRow>
-                );
-              })
-            }
-          </TableBody>
-        </Table>
+        <ActivityTable
+          activities={this.state.activities}
+          categories={this.state.categories}
+          onDelete={this.deleteActivity}
+          onChangeCategory={this.onChange}
+          onToggleActivity={this.toggleCheckbox}
+        />
       </FlexContainerCol>
     );
   }
